Closer data type

+++ src/assets.rs
@@ -6,7 +6,7 @@ pub struct Assets {
 #[cfg(test)]
 impl Assets {
     #[cfg(test)]
-    pub fn get_player_filename(&self, player_index: i8, singer_index: i8) -> String {
+    pub fn get_player_filename(&self, player_index: usize, singer_index: usize) -> String {
         let band = crate::band::Band{};
         let mut filename = band.get_name(player_index, singer_index);
         filename.push_str(".png");
